Move model price sync apply logic into module
